rename: 모달 테스트 페이지 파일명 수정(#47)
