+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Route, Routes } from 'react-router-dom'
 import './App.css'
 import MainLayout from '@layouts/MainLayout'
 import TestHub from '@src/pages/test-page/TestHub'
-import TestApplicationModalPage from './pages/test-page/TestApplicationModalPage'
+import TestApplicationModalPage from './pages/test-page/TestModalPage'
 import { ROUTES } from './constants/routes'
 import RecruitmentPase from '@src/pages/recruitment-page/RecruitmentPage'
 import TestUIPage from './pages/test-page/TestUIPage'
